refactor: restaurant.ts 로직별 파일 분리
