+++ src/components/modal/restaurant-detail-modal.ts
@@ -1,9 +1,6 @@
 import $restaurantDetailContent from "../restaurant/restaurant-detail-content.ts";
-import {
-  Restaurant,
-  currentRestaurantData,
-  saveRestaurantsToLocalStorage,
-} from "../../data/restaurant.ts";
+import { Restaurant } from "../../data/models/restaurant.ts";
+import { currentRestaurantData, saveRestaurantsToLocalStorage } from "../../data/storage/restaurantStorage.ts";
 
 export const handleRestaurantDetailModalClose = (): void => {
   const modal = document.querySelector(".restaurant-detail-modal");
