+++ src/data/models/restaurant.ts
@@ -1,0 +1,11 @@
+export type Restaurant = {
+  dataId: number;
+  categoryIcon: string;
+  categoryTitle: string;
+  name: string;
+  distance: number;
+  distanceCaption: string;
+  description: string;
+  link: string;
+  isFavorite: boolean;
+};
